requirements.txt
# OpenWakeWord Trainer dependencies.
# torch / torchaudio are installed separately from the cu128 index in the Dockerfile.

# Core training
openwakeword==0.6.0
onnx==1.17.0
onnxruntime-gpu==1.20.1

# TTS
piper-tts==1.4.2
elevenlabs==2.46.0

# Audio I/O + augmentation
audiomentations==0.43.1
soundfile==0.12.1
librosa==0.10.2.post1
pydub==0.25.1
numpy==1.26.4
scipy==1.13.1

# Datasets / downloads
huggingface_hub==0.27.0
datasets==3.2.0
requests==2.32.3
tqdm==4.67.1

# Web UI
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.20
sse-starlette==2.1.3
pydantic==2.10.4
pydantic-settings==2.7.0

# Utilities
PyYAML==6.0.2
torchinfo==1.8.0
torchmetrics==1.6.1
mutagen==1.47.0
acoustics==0.2.6
